+++ src/main/java/backend/objects/FastActingInsulin.java
@@ -3,12 +3,12 @@ package backend.objects;
 public class FastActingInsulin extends Injections {
 
     //note: times and dosage MUST be parallel arrays
-    public FastActingInsulin(String[] times, int[] dosage) {
+    public FastActingInsulin(String[] times, double[] dosage) {
         super(times, dosage);
     }
 
     @Override
-    public int[][] getInjectionsArr() {
+    public double[][] getInjectionsArr() {
         return super.getInjectionsArr();
     }
 }
